delete mm.h and mm.cpp
